requirements.txt
# 项目依赖
torch>=2.0.0
dgl>=1.1.2
opencv-python>=4.7.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
pillow>=9.4.0
pandas>=2.0.0
tqdm>=4.67.0
networkx>=3.1.0
flask>=2.0.0
scikit-image>=0.19.0
scikit-learn>=1.0.0
pytest>=7.0.0
tensorboard>=2.10.0